solution: test_SnakeGame01 test_initial_snake_position
In Python 3 the `//` operator is used to divide a number and discard the
fractional part. It can be used to calculate the initial position of the
snake because it will take care of the case when the provided width or
height is odd.
The initial position should be set only if the `snake` argument is not
provided (note that the previous test will fail if this is not
guaranteed). This can be done by using an `if` statement like this:
```
if snake == None:
    self.snake = [(width // 2, height // 2)]
else:
    self.snake = snake
```
It is possible to write this in a shorter form by using the `or`
operator:
```
self.snake = snake or [(width // 2, height // 2)]
```