solution: test_SnakeGame02 test_move_snake_head
The snake only consists of the head right now. To move snake's head we
need first to read its current position. Snake's head position is always
the first item in the `self.snake` list, so we can access it using 0 as
the index (i.e. `self.snake[0]`). The returned value will be a pair
containing head's coordinates. This pair cannot be changed so we have to
extract the x and y coordinates. We can do it as follows:
```
x, y = self.snake[0]
```
The code above extracts the head of the snake from the list and sets the
`x` and `y` variables to the respective pair values.
Once the coordinates has been extracted we need to adjust them based on
the direction. We need to increase or decrease the `x` coordinate if the
snake is moving to the right or left. Similary, we need to increase or
decrease the `y` coordinate if the snake is moving down or up.
After calculating new coordinates we need to replace the current list
representing the snake with a new list containing new coordinates. We
can do it because currently the snake consists only of the head.