solution: test_SnakeGame02 test_change_direction_added
We need to add a method that allows changing snake's direction. The new
direction will be passed in the `new_direction` parameter. The first
question to ask is if we always should change the direction. For
instance if the current direction is UP and the new direction is UP
there is no change required (or the change will not have any effect).
Now, what if the current direction is UP and the new direction is DOWN?
In that case the direction should not change because the snake would
immediately eat itself. Similarly when the snake is moving sideways,
trying to chang the direction to LEFT or RIGHT should have no effect.
In other cases we can just update the direction to new direction passed
to the method.
This could be implemented by checking all the conditions as follows:
```
if (self.direction == UP or self.direction == DOWN) and (new_direction == UP or new_direction == DOWN):
    return
if (self.direction == LEFT or self.direction == RIGHT) and (new_direction == LEFT or new_direction == RIGHT):
    return
self.direction = new_direction
```
but we could also do some math to simplify it. UP, DOWN, LEFT and RIGHT
variables are assigned numbers. UP and DOWN are even and LEFT and RIGHT
are odd. We only need to change the direction if the parity of the new
direction is different from the current direction. We can check the
parity by using the `%` operator which returns the remainder of the
division - we simply need to compare remainders of dividing the current
and new direction by 2 and update the current direction if the values
are different. This is how we end up with:
```
if new_direction % 2 != self.direction % 2:
    self.direction = new_direction
```